NFC_NCIHALx_AR18C0.8.2.9_O_OpnSrc

+++ external/libese-spi/p73/pal/phNxpEsePal.c
@@ -223,7 +223,7 @@ void phPalEse_print_packet(const char *pString, const uint8_t *p_data,
 ** Returns          None
 **
 *******************************************************************************/
-void phPalEse_sleep(uint32_t usec)
+void phPalEse_sleep(long usec)
 {
     usleep(usec);
     return;
